Rename solution/project and clean up code
