+++ ChromaTemp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27130.2010
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RazerTempREST.App", "RazerTempREST.App\RazerTempREST.App.csproj", "{ED94A3EF-4D53-4E75-8A7D-EC4BE4C11876}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChromaTemp", "ChromaTemp\ChromaTemp.csproj", "{ED94A3EF-4D53-4E75-8A7D-EC4BE4C11876}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
