+++ ChromaTemp/ChromaTemp.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{ED94A3EF-4D53-4E75-8A7D-EC4BE4C11876}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>RazerTempREST.App</RootNamespace>
-    <AssemblyName>RazerTempREST.App</AssemblyName>
+    <RootNamespace>ChromaTemp</RootNamespace>
+    <AssemblyName>ChromaTemp</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -40,11 +40,6 @@
     </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
